refactor:ipsec: move IPSecParameters under pkg/datapath/types
This commit moves IPSecParameters to pkg/datapath/types. This is needed, as in
subsequent commits few methods from the IPSecAgent that accepts such structure
as parameter will be moved in the interface. Therefore, let's declare such
data in the same place.
