+++ pkg/datapath/linux/ipsec.go
@@ -18,6 +18,7 @@ import (
 	"github.com/cilium/cilium/pkg/datapath/linux/linux_defaults"
 	"github.com/cilium/cilium/pkg/datapath/linux/route"
 	"github.com/cilium/cilium/pkg/datapath/linux/safenetlink"
+	"github.com/cilium/cilium/pkg/datapath/types"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/metrics"
 	"github.com/cilium/cilium/pkg/node"
@@ -201,7 +202,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4DoSubnetEncryption(newNode *nodeTypes.
 	}
 
 	// The common bits which are consistent between XFRM policy/state creation.
-	template := &ipsec.IPSecParameters{
+	template := &types.IPSecParameters{
 		LocalBootID:    node.GetBootID(n.log),
 		RemoteBootID:   newNode.BootID,
 		RemoteNodeID:   nodeID,
@@ -211,7 +212,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4DoSubnetEncryption(newNode *nodeTypes.
 	}
 
 	for _, cidr := range n.nodeConfig.GetIPv4PodSubnets() {
-		params := ipsec.NewIPSecParamaters(template)
+		params := types.NewIPSecParameters(template)
 		params.Dir = ipsec.IPSecDirOut
 		params.SourceSubnet = wildcardCIDR
 		params.DestSubnet = cidr
@@ -225,7 +226,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4DoSubnetEncryption(newNode *nodeTypes.
 	}
 
 	// insert fwd rule
-	params := ipsec.NewIPSecParamaters(template)
+	params := types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirFwd
 	params.SourceSubnet = wildcardCIDR
 	params.DestSubnet = wildcardCIDR
@@ -234,7 +235,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4DoSubnetEncryption(newNode *nodeTypes.
 	spi, err = ipsec.UpsertIPsecEndpoint(n.log, params)
 	errs = errors.Join(errs, upsertIPsecLog(n.log, err, "fwd IPv4", params.SourceSubnet, params.DestSubnet, spi, nodeID))
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirIn
 	params.SourceSubnet = wildcardCIDR
 	params.DestSubnet = wildcardCIDR
@@ -292,7 +293,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4Do(oldNode, newNode *nodeTypes.Node, n
 	}
 
 	// The common bits which are consistent between XFRM policy/state creation.
-	template := &ipsec.IPSecParameters{
+	template := &types.IPSecParameters{
 		LocalBootID:    node.GetBootID(n.log),
 		RemoteBootID:   newNode.BootID,
 		RemoteNodeID:   nodeID,
@@ -305,7 +306,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4Do(oldNode, newNode *nodeTypes.Node, n
 	// since some other ipsec related parameters beside the remote CIDR might have been changed
 	// (e.g: RemoteBootID, RemoteNodeID and so on)
 	for _, remoteCIDR := range cidr.CIDRsToIPNets(newNode.GetIPv4AllocCIDRs()) {
-		params := ipsec.NewIPSecParamaters(template)
+		params := types.NewIPSecParameters(template)
 		params.Dir = ipsec.IPSecDirOut
 		params.SourceSubnet = wildcardCIDR
 		params.DestSubnet = remoteCIDR
@@ -334,7 +335,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4Do(oldNode, newNode *nodeTypes.Node, n
 	}
 
 	// insert fwd rule
-	params := ipsec.NewIPSecParamaters(template)
+	params := types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirFwd
 	params.SourceSubnet = wildcardCIDR
 	params.DestSubnet = wildcardCIDR
@@ -343,7 +344,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4Do(oldNode, newNode *nodeTypes.Node, n
 	spi, err = ipsec.UpsertIPsecEndpoint(n.log, params)
 	errs = errors.Join(errs, upsertIPsecLog(n.log, err, "fwd IPv4", params.SourceSubnet, params.DestSubnet, spi, nodeID))
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirIn
 	params.SourceSubnet = wildcardCIDR
 	params.DestSubnet = wildcardCIDR
@@ -373,7 +374,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4Do(oldNode, newNode *nodeTypes.Node, n
 	localOverlayIPExactMatch := &net.IPNet{IP: localUnderlayIP, Mask: exactMatchMaskIPv4}
 	remoteOverlayIPExactMatch := &net.IPNet{IP: remoteUnderlayIP, Mask: exactMatchMaskIPv4}
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirOut
 	params.SourceSubnet = localOverlayIPExactMatch
 	params.DestSubnet = remoteOverlayIPExactMatch
@@ -385,7 +386,7 @@ func (n *linuxNodeHandler) enableIPSecIPv4Do(oldNode, newNode *nodeTypes.Node, n
 		statesUpdated = false
 	}
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirIn
 	params.SourceSubnet = wildcardCIDR
 	params.DestSubnet = wildcardCIDR
@@ -480,7 +481,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6DoSubnetEncryption(newNode *nodeTypes.
 	}
 
 	// The common bits which are consistent between XFRM policy/state creation.
-	template := &ipsec.IPSecParameters{
+	template := &types.IPSecParameters{
 		LocalBootID:    node.GetBootID(n.log),
 		RemoteBootID:   newNode.BootID,
 		RemoteNodeID:   nodeID,
@@ -490,7 +491,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6DoSubnetEncryption(newNode *nodeTypes.
 	}
 
 	for _, cidr := range n.nodeConfig.GetIPv6PodSubnets() {
-		params := ipsec.NewIPSecParamaters(template)
+		params := types.NewIPSecParameters(template)
 		params.Dir = ipsec.IPSecDirOut
 		params.SourceSubnet = wildcardCIDR6
 		params.DestSubnet = cidr
@@ -503,7 +504,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6DoSubnetEncryption(newNode *nodeTypes.
 		}
 	}
 
-	params := ipsec.NewIPSecParamaters(template)
+	params := types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirFwd
 	params.SourceSubnet = wildcardCIDR6
 	params.DestSubnet = wildcardCIDR6
@@ -515,7 +516,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6DoSubnetEncryption(newNode *nodeTypes.
 		statesUpdated = false
 	}
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirIn
 	params.SourceSubnet = wildcardCIDR6
 	params.DestSubnet = wildcardCIDR6
@@ -571,7 +572,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6Do(oldNode, newNode *nodeTypes.Node, n
 	}
 
 	// The common bits which are consistent between XFRM policy/state creation.
-	template := &ipsec.IPSecParameters{
+	template := &types.IPSecParameters{
 		LocalBootID:    node.GetBootID(n.log),
 		RemoteBootID:   newNode.BootID,
 		RemoteNodeID:   nodeID,
@@ -584,7 +585,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6Do(oldNode, newNode *nodeTypes.Node, n
 	// since some other ipsec related parameters beside the remote CIDR might have been changed
 	// (e.g: RemoteBootID, RemoteNodeID and so on)
 	for _, remoteCIDR := range cidr.CIDRsToIPNets(newNode.GetIPv6AllocCIDRs()) {
-		params := ipsec.NewIPSecParamaters(template)
+		params := types.NewIPSecParameters(template)
 		params.Dir = ipsec.IPSecDirOut
 		params.SourceSubnet = wildcardCIDR6
 		params.DestSubnet = remoteCIDR
@@ -613,7 +614,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6Do(oldNode, newNode *nodeTypes.Node, n
 	}
 
 	// insert forward policy
-	params := ipsec.NewIPSecParamaters(template)
+	params := types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirFwd
 	params.SourceSubnet = wildcardCIDR6
 	params.DestSubnet = wildcardCIDR6
@@ -625,7 +626,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6Do(oldNode, newNode *nodeTypes.Node, n
 		statesUpdated = false
 	}
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirIn
 	params.SourceSubnet = wildcardCIDR6
 	params.DestSubnet = wildcardCIDR6
@@ -655,7 +656,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6Do(oldNode, newNode *nodeTypes.Node, n
 	localOverlayIPExactMatch := &net.IPNet{IP: localUnderlayIP, Mask: exactMatchMaskIPv6}
 	remoteOverlayIPExactMatch := &net.IPNet{IP: remoteUnderlayIP, Mask: exactMatchMaskIPv6}
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirOut
 	params.SourceSubnet = localOverlayIPExactMatch
 	params.DestSubnet = remoteOverlayIPExactMatch
@@ -667,7 +668,7 @@ func (n *linuxNodeHandler) enableIPSecIPv6Do(oldNode, newNode *nodeTypes.Node, n
 		statesUpdated = false
 	}
 
-	params = ipsec.NewIPSecParamaters(template)
+	params = types.NewIPSecParameters(template)
 	params.Dir = ipsec.IPSecDirIn
 	params.SourceSubnet = wildcardCIDR
 	params.DestSubnet = wildcardCIDR
