+++ pkg/datapath/linux/ipsec/ipsec_linux_test.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/cilium/cilium/pkg/datapath/linux/linux_defaults"
 	"github.com/cilium/cilium/pkg/datapath/linux/safenetlink"
+	"github.com/cilium/cilium/pkg/datapath/types"
 	"github.com/cilium/cilium/pkg/node"
 	"github.com/cilium/cilium/pkg/testutils"
 )
@@ -120,7 +121,7 @@ func testInvalidLoadKeys(t *testing.T) {
 	_, _, err := LoadIPSecKeys(keys)
 	require.Error(t, err)
 
-	params := &IPSecParameters{
+	params := &types.IPSecParameters{
 		LocalBootID:    localBootID,
 		RemoteBootID:   remoteBootID,
 		RemoteNodeID:   remoteNodeID,
@@ -219,7 +220,7 @@ func testUpsertIPSecEquals(t *testing.T) {
 	ipSecKeysGlobal[remote.IP.String()] = key
 	ipSecKeysGlobal[""] = key
 
-	params := &IPSecParameters{
+	params := &types.IPSecParameters{
 		LocalBootID:    localBootID,
 		RemoteBootID:   remoteBootID,
 		RemoteNodeID:   remoteNodeID,
@@ -296,7 +297,7 @@ func testUpsertIPSecEndpointOut(t *testing.T) {
 	ipSecKeysGlobal[remote.IP.String()] = key
 	ipSecKeysGlobal[""] = key
 
-	params := &IPSecParameters{
+	params := &types.IPSecParameters{
 		LocalBootID:    localBootID,
 		RemoteBootID:   remoteBootID,
 		RemoteNodeID:   0xBEEF,
@@ -410,7 +411,7 @@ func testUpsertIPSecEndpointFwd(t *testing.T) {
 	ipSecKeysGlobal[remote.IP.String()] = key
 	ipSecKeysGlobal[""] = key
 
-	params := &IPSecParameters{
+	params := &types.IPSecParameters{
 		LocalBootID:    localBootID,
 		RemoteBootID:   remoteBootID,
 		RemoteNodeID:   0xBEEF,
@@ -503,7 +504,7 @@ func testUpsertIPSecEndpointIn(t *testing.T) {
 	ipSecKeysGlobal[remote.IP.String()] = key
 	ipSecKeysGlobal[""] = key
 
-	params := &IPSecParameters{
+	params := &types.IPSecParameters{
 		LocalBootID:    localBootID,
 		RemoteBootID:   remoteBootID,
 		RemoteNodeID:   0xBEEF,
@@ -586,7 +587,7 @@ func testUpsertIPSecEndpointIn(t *testing.T) {
 }
 
 func testUpsertIPSecKeyMissing(t *testing.T) {
-	params := &IPSecParameters{
+	params := &types.IPSecParameters{
 		LocalBootID:    localBootID,
 		RemoteBootID:   remoteBootID,
 		RemoteNodeID:   remoteNodeID,
@@ -620,7 +621,7 @@ func testUpdateExistingIPSecEndpoint(t *testing.T) {
 	ipSecKeysGlobal[remote.IP.String()] = key
 	ipSecKeysGlobal[""] = key
 
-	params := &IPSecParameters{
+	params := &types.IPSecParameters{
 		LocalBootID:    localBootID,
 		RemoteBootID:   remoteBootID,
 		RemoteNodeID:   0xBEEF,
